added rtc.c, added real time spectrum display to chart,added real time cps cpm data

+++ FW/Core/Src/communication.c
@@ -8,10 +8,10 @@
 #include "communication.h"
 
 
-uint32_t PrintToBuffer(uint16_t data[],uint16_t size){
+uint32_t PrintToBuffer(uint32_t data[],uint16_t size){
 	uint32_t index = 0;
 	for (uint32_t i = 0;i < size;i++){
-		index = index + sprintf(UserTxBufferFS + index,"%d:%d\r\n",i, data[i]);
+		index = index + sprintf(UserTxBufferFS + index,"%d:%lu\r\n",i, data[i]);
 	}
 	return index;
 }
